+++ lib/src/controls/form/pickers/pickers.dart
@@ -91,7 +91,7 @@ class YesNoPickerControl extends StatelessWidget {
   Widget build(BuildContext context) {
     assert(debugCheckHasFluentTheme(context));
 
-    ButtonStyle buttonStyle = ButtonStyle(
+    final buttonStyle = ButtonStyle(
       elevation: ButtonState.all(0.0),
       backgroundColor: ButtonState.resolveWith(
         (states) => ButtonThemeData.uncheckedInputColor(
@@ -273,7 +273,7 @@ extension FixedExtentScrollControllerExtension on FixedExtentScrollController {
 
     if (forward) {
       final currentItem = selectedItem;
-      int to = currentItem + 1;
+      var to = currentItem + 1;
       if (currentItem == amount - 1) to = 0;
 
       return animateToItem(
@@ -283,7 +283,7 @@ extension FixedExtentScrollControllerExtension on FixedExtentScrollController {
       );
     } else {
       final currentItem = selectedItem;
-      int to = currentItem - 1;
+      var to = currentItem - 1;
       if (currentItem == 0) to = amount - 1;
 
       return animateToItem(
@@ -318,7 +318,7 @@ class Picker extends StatefulWidget {
 }
 
 class _PickerState extends State<Picker> {
-  final GlobalKey _childKey = GlobalKey();
+  late final GlobalKey _childKey = GlobalKey(debugLabel: '${widget.child} key');
 
   Future<void> open() {
     assert(
@@ -330,6 +330,7 @@ class _PickerState extends State<Picker> {
 
     final navigator = Navigator.of(context);
     final isAcrylicDisabled = DisableAcrylic.of(context) != null;
+
     return navigator.push(PageRouteBuilder(
       barrierColor: Colors.transparent,
       opaque: false,
@@ -345,7 +346,7 @@ class _PickerState extends State<Picker> {
         // value from the Win UI 3 Gallery
         final centeredOffset = widget.pickerHeight * 0.41;
         // the popup menu y is the [button y] - [y of highlight tile]
-        double y = childOffset.dy - centeredOffset;
+        var y = childOffset.dy - centeredOffset;
 
         // if the popup menu [y] + picker height overlaps the screen height, make
         // it to the bottom of the screen
@@ -371,8 +372,10 @@ class _PickerState extends State<Picker> {
         }();
 
         final view = Stack(children: [
-          PositionedDirectional(
-            start: x,
+          // We can not use PositionedDirectional here
+          // See https://github.com/bdlukaa/fluent_ui/issues/675
+          Positioned(
+            left: x,
             top: y,
             height: widget.pickerHeight,
             width: width,