+++ lib/src/controls/navigation/navigation_view/body.dart
@@ -104,7 +104,7 @@ class _NavigationBodyState extends State<_NavigationBody> {
     final view = InheritedNavigationView.of(context);
     final theme = FluentTheme.of(context);
 
-    return Container(
+    return ColoredBox(
       color: theme.scaffoldBackgroundColor,
       child: AnimatedSwitcher(
         switchInCurve: widget.animationCurve ?? Curves.ease,
@@ -120,7 +120,7 @@ class _NavigationBodyState extends State<_NavigationBody> {
             return widget.transitionBuilder!(child, animation);
           }
 
-          bool isTop = view.displayMode == PaneDisplayMode.top;
+          var isTop = view.displayMode == PaneDisplayMode.top;
 
           if (isTop) {
             // Other transtitions other than default is only applied to top nav
@@ -137,32 +137,41 @@ class _NavigationBodyState extends State<_NavigationBody> {
 
           return EntrancePageTransition(
             animation: animation,
-            vertical: true,
             child: child,
           );
         },
-        child: KeyedSubtree(
-          key: widget.itemKey,
-          child: PageView.builder(
-            key: _pageKey,
-            physics: const NeverScrollableScrollPhysics(),
-            allowImplicitScrolling: false,
-            controller: pageController,
-            itemCount: view.pane!.effectiveItems.length,
-            itemBuilder: (context, index) {
-              final bool isSelected = view.pane!.selected == index;
-              final item = view.pane!.effectiveItems[index];
+        child: () {
+          final paneBodyBuilder = widget.paneBodyBuilder;
+          if (paneBodyBuilder != null) {
+            return FocusTraversalGroup(
+              child: paneBodyBuilder.call(view.pane?.selected != null
+                  ? view.pane?.selectedItem.body
+                  : null),
+            );
+          } else {
+            return KeyedSubtree(
+              key: widget.itemKey,
+              child: PageView.builder(
+                key: _pageKey,
+                physics: const NeverScrollableScrollPhysics(),
+                controller: pageController,
+                itemCount: view.pane!.effectiveItems.length,
+                itemBuilder: (context, index) {
+                  final isSelected = view.pane!.selected == index;
+                  final item = view.pane!.effectiveItems[index];
 
-              return ExcludeFocus(
-                key: item.bodyKey,
-                excluding: !isSelected,
-                child: FocusTraversalGroup(
-                  child: widget.paneBodyBuilder?.call(item.body) ?? item.body,
-                ),
-              );
-            },
-          ),
-        ),
+                  return ExcludeFocus(
+                    key: item.bodyKey,
+                    excluding: !isSelected,
+                    child: FocusTraversalGroup(
+                      child: item.body,
+                    ),
+                  );
+                },
+              ),
+            );
+          }
+        }(),
       ),
     );
   }