+++ lib/src/controls/navigation/navigation_view/view.dart
@@ -64,6 +64,10 @@ class NavigationView extends StatefulWidget {
   /// and the body of the navigation pane is dynamically determined or
   /// affected by the current route rather than just by the currently
   /// selected pane.
+  ///
+  /// If this is not null then this builder will be responsible for state
+  /// management of the child widget. One way to accomplish this is to
+  /// use an [IndexedStack].
   final NavigationContentBuilder? paneBodyBuilder;
 
   /// The navigation pane, that can be displayed either on the
@@ -131,7 +135,7 @@ class NavigationView extends StatefulWidget {
   }
 
   @override
-  NavigationViewState createState() => NavigationViewState();
+  State<NavigationView> createState() => NavigationViewState();
 }
 
 class NavigationViewState extends State<NavigationView> {
@@ -187,7 +191,6 @@ class NavigationViewState extends State<NavigationView> {
     paneScrollController = widget.pane?.scrollController ??
         ScrollController(
           debugLabel: '${widget.runtimeType} scroll controller',
-          keepScrollOffset: true,
         );
     paneScrollController.addListener(_handleScrollControllerEvent);
 
@@ -215,7 +218,9 @@ class NavigationViewState extends State<NavigationView> {
     if (oldWidget.pane?.selected != widget.pane?.selected) {
       _oldIndex = oldWidget.pane?.selected ?? -1;
 
-      final item = widget.pane?.selectedItem.itemKey.currentContext;
+      final item = widget.pane?.selected == null
+          ? null
+          : widget.pane?.selectedItem.itemKey.currentContext;
 
       if (item != null) {
         final atEnd =
@@ -281,14 +286,14 @@ class NavigationViewState extends State<NavigationView> {
     assert(debugCheckHasMediaQuery(context));
     assert(debugCheckHasDirectionality(context));
 
-    final Brightness brightness = FluentTheme.of(context).brightness;
-    final NavigationPaneThemeData theme = NavigationPaneTheme.of(context);
-    final FluentLocalizations localizations = FluentLocalizations.of(context);
-    final MediaQueryData mediaQuery = MediaQuery.of(context);
+    final brightness = FluentTheme.of(context).brightness;
+    final theme = NavigationPaneTheme.of(context);
+    final localizations = FluentLocalizations.of(context);
+    final mediaQuery = MediaQuery.of(context);
     final EdgeInsetsGeometry appBarPadding = EdgeInsetsDirectional.only(
       top: widget.appBar?.finalHeight(context) ?? 0.0,
     );
-    final TextDirection direction = Directionality.of(context);
+    final direction = Directionality.of(context);
 
     Color? overlayBackgroundColor() {
       if (theme.backgroundColor == null) {
@@ -337,7 +342,7 @@ class NavigationViewState extends State<NavigationView> {
         /// - A left, icon-only, nav pane (compact) on medium window widths
         /// (641px to 1007px).
         /// - Only a menu button (minimal) on small window widths (640px or less).
-        double width = consts.biggest.width;
+        var width = consts.biggest.width;
         if (width.isInfinite) width = mediaQuery.size.width;
 
         if (width <= 640) {
@@ -352,7 +357,7 @@ class NavigationViewState extends State<NavigationView> {
       }
       assert(displayMode != PaneDisplayMode.auto);
 
-      Widget appBar = () {
+      var appBar = () {
         if (widget.appBar != null) {
           return _NavigationAppBar(
             appBar: widget.appBar!,
@@ -458,11 +463,11 @@ class NavigationViewState extends State<NavigationView> {
                   ) as bool? ??
                   _compactOverlayOpen;
 
-              double openSize =
+              var openSize =
                   pane.size?.openPaneWidth ?? kOpenNavigationPaneWidth;
 
-              final bool noOverlayRequired = consts.maxWidth / 2.5 > openSize;
-              final bool openedWithoutOverlay =
+              final noOverlayRequired = consts.maxWidth / 2.5 > openSize;
+              final openedWithoutOverlay =
                   _compactOverlayOpen && consts.maxWidth / 2.5 > openSize;
 
               // print(
@@ -637,18 +642,18 @@ class NavigationViewState extends State<NavigationView> {
                         border: Border.all(
                           color: const Color(0xFF6c6c6c),
                           width: 0.15,
+                            ),
+                            borderRadius: BorderRadius.circular(8.0),
+                          ),
+                          margin: const EdgeInsets.symmetric(vertical: 1.0),
+                          padding: appBarPadding,
+                          child: _OpenNavigationPane(
+                            theme: theme,
+                            pane: pane,
+                            paneKey: _panelKey,
+                            listKey: _listKey,
+                            onItemSelected: () => minimalPaneOpen = false,
                         ),
-                        borderRadius: BorderRadius.circular(8.0),
-                      ),
-                      margin: const EdgeInsets.symmetric(vertical: 1.0),
-                      padding: appBarPadding,
-                      child: _OpenNavigationPane(
-                        theme: theme,
-                        pane: pane,
-                        paneKey: _panelKey,
-                        listKey: _listKey,
-                        onItemSelected: () => minimalPaneOpen = false,
-                      ),
                     ),
                   ),
                 ),
@@ -768,7 +773,7 @@ class NavigationAppBar with Diagnosticable {
         widget = leading!;
       } else if (automaticallyImplyLeading && imply) {
         final ModalRoute<dynamic>? parentRoute = ModalRoute.of(context);
-        final bool canPop = parentRoute?.canPop ?? false;
+        final canPop = parentRoute?.canPop ?? false;
 
         assert(debugCheckHasFluentLocalizations(context));
         assert(debugCheckHasFluentTheme(context));
@@ -833,9 +838,8 @@ class _NavigationAppBar extends StatelessWidget {
 
     final mediaQuery = MediaQuery.of(context);
 
-    final PaneDisplayMode displayMode =
-        InheritedNavigationView.maybeOf(context)?.displayMode ??
-            PaneDisplayMode.top;
+    final displayMode = InheritedNavigationView.maybeOf(context)?.displayMode ??
+        PaneDisplayMode.top;
     final leading = appBar._buildLeading(displayMode != PaneDisplayMode.top);
     final title = () {
       if (appBar.title != null) {
@@ -851,7 +855,6 @@ class _NavigationAppBar extends StatelessWidget {
           child: DefaultTextStyle(
             style:
                 FluentTheme.of(context).typography.caption ?? const TextStyle(),
-            overflow: TextOverflow.clip,
             maxLines: 1,
             softWrap: false,
             child: appBar.title!,